src/requirements.txt
celery==5.3.6
Django==4.2.5
django-autoslug==1.9.9
django-ckeditor==6.7.0
django_jazzmin==2.6.0
django-cleanup==8.0.0
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
pillow==10.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
redis==5.0.1